requirements.txt
flask
python-telegram-bot
python-dotenv
gunicorn
pymongo[srv]
